third_party/WebKit/LayoutTests/fast/dom/shadow/multiple-host-pseudos-in-compound-expected.txt

CONSOLE WARNING: line 7: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Multiple :host(-context) pseudos in same compound.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(host).color is "rgb(0, 128, 0)"
PASS getComputedStyle(host, '::before').color is "rgb(0, 128, 0)"
PASS getComputedStyle(host).color is "rgb(0, 128, 0)"
PASS getComputedStyle(host, '::before').color is "rgb(0, 128, 0)"
PASS getComputedStyle(host).color is "rgb(0, 128, 0)"
PASS getComputedStyle(host, '::before').color is "rgb(0, 128, 0)"
PASS getComputedStyle(host).color is "rgb(0, 128, 0)"
PASS getComputedStyle(host, '::before').color is "rgb(0, 128, 0)"
PASS getComputedStyle(host).color is "rgb(0, 128, 0)"
PASS getComputedStyle(host, '::before').color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE

Should be green
